Changed how the game deals with the edge of levels
Made objects move to the oppostie side when they get too far away from the Player (or adverage position of the Players) instead of having a fixed level area where the objects move to the opposite side of the level when the move off the edge. 

This makes the game seem smooter to the player, since he will never move off the edge (in single player) and the objects that do move off the edge to the other side will be done far away from him.