Different Aspect Ratios, Static Level Option, Custom Loss/Win Conditions
Added support for different aspect ratios. If aspect ratio is different then levelsize ratio, even at largest zoom, all of the level will not be shown at once. 

Added an option for objects to be moved to the opposite side the old way or the new way. 

Added virtual methods in Level for levels to override so they can add their own custom win and loss conditions. 

Changed and fixed some other things.