Finished Optons, Started on Audio
Finished setting up the Options menus. Seems to work but needs tested more. 

Resolution and fullscreen cant be changed in the editor. So I have not tested those yet. 

Started working on setting up a way to have audio work in a unified way, so the volume options will work.