Fixed Bugs with Login, CRUD , Asteroid and Blob
Fixed bug making it so you could not log in.
Fixed bug making logged in accounts treated as trail accounts.
Fix bug in Asteroid and Blob that caused them to breakup twice if they got damaged twice in the same frame. This caused them to create too many pieces.

Redid these commits, which I did a few days ago, because I accidenlty committed the executable I made for this game when I compiled it, which caused my local repository to become large.